Bevy 0.18 -> 0.19, and rustc 1.92 -> 1.97
Fifth and last step. Bevy 0.19 requires rustc 1.95, so the toolchain moved to
1.97.1 and the workspace MSRV from 1.82 to 1.95 — a thirteen-version jump, taken
deliberately rather than drifted into.

The change is egui's, not Bevy's. `SidePanel` and `TopBottomPanel` are one
`Panel` now, and a panel is shown **inside a `Ui`** rather than against a
`Context` — so the viewport becomes an explicit root `Ui` on the background
layer and the whole layout hangs off that. It is a better model than the one it
replaces: a panel was always a region of something, and now it says so. What was
`ctx.available_rect()` is `root.available_rect_before_wrap()`, which is the same
question asked of the thing that actually knows the answer.

Renames alongside: `default_width`/`default_height` -> `default_size`,
`width_range`/`height_range` -> `size_range`, `wants_pointer_input` ->
`egui_wants_pointer_input`, `is_pointer_over_area` -> `is_pointer_over_egui`,
`wants_keyboard_input` -> `egui_wants_keyboard_input`. Asset handles from
`Assets::get_mut` need `mut` bindings.

Five new lints from the newer compiler, all in mm-core and all real:
`is_multiple_of` in four places and one `as i64` on something already `i64`.

The upgrade in total, 0.14.2 -> 0.19.0 across five commits: **one file**.
`main.rs` is the only thing in this crate that knows Bevy exists, and all 154
mm-app library tests passed untouched through every step because none of them
can see a renderer. That is the wall in `slide.rs` earning its keep, and it is
the strongest argument yet for leaving it exactly where it is.

It also vindicates doing the SDF renderer *before* the upgrade rather than
after. `Material2d`, `MeshVertexAttribute` and `specialize` are the same API at
0.19 that they were at 0.14, and `cell.wgsl` needed no changes at all — so the
work ported for the cost of some import paths and a two-line spawn, which is
what `docs/UI.md` §7 predicted and now records per version.

Verified on 0.19: renders correctly at 1400x — cells, organelles, chemical field
and the full interface — at 139 fps and 59 t/s. 298 mm-core lib, 154 mm-app lib,
39 mm-asm/mm-cli; every mm-core integration binary green (m8_ecology 932s,
m3_chemotaxis 684s, m2_life 385s, m5_phylogeny 155s); 6 m4_microscope, 8
m6_tools; clippy clean across the workspace.

+++ crates/mm-core/src/sensing.rs
@@ -373,7 +373,7 @@ pub fn step_physics(
         if ny != cells.y[i] as i64 {
             cells.vy[i] = 0;
         }
-        report.moved += (step_x.abs() + step_y.abs()) as i64;
+        report.moved += step_x.abs() + step_y.abs();
 
         // --- the reaction: what the cilia pushed against ---
         //
